backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==10.4.0
numpy==2.1.3

# Py3.13 组合（若网络原因失败，可改成 2.9.* / 0.20.* 让 pip 自动解）
torch==2.9.0
torchvision==0.23.0

ultralytics==8.3.50

# 如不需视频/绘图，删除这行可减小体积
opencv-python-headless==4.10.0.84
